reorg
Fixing my garbage coding practices
